(2.2x) backport r11881 for #4019 "kill useless casts" -- patch from jlouis

+++ libtransmission/fdlimit.c
@@ -793,7 +793,7 @@ tr_fdSetPeerLimit( tr_session * session, int socket_limit )
 #endif
     gFd->public_socket_limit = socket_limit;
 
-    tr_dbg( "socket limit is %d", (int)gFd->socket_limit );
+    tr_dbg( "socket limit is %d", gFd->socket_limit );
 }
 
 int
